don't check for a CF challenge, looks like we're automatically bypassing it again.
TODO: Add a quick, proper check for a CF challenge on each new page load. This could be a new `wait_for_page_load(check_for_cf_challenge=True)` or similar.

Right now the check for the challenge takes a long time so we don't want to call it and waste a bunch of time if we don't need it.