try to get unblocked when setting up 2fa
Current commit gets a cloudflare challenge 100% of the time when going to the account management page to setup 2fa. This could be happening for multiple reasons..

Changed:
* Update brave browser download to the latest version.
* Update the user agent to the latest version.
* Randomize typing interval for each keystroke.
* Wait a few seconds before going to the management page after initial account creation.

This worked for me but you'll still get blocked after a while from using the same IP over and over again, since Brave is Chromium based, and Chrome took away to ability to easily load our local proxy extension.